+++ packages/ui/src/components/Input/Input/index.ts
@@ -1,5 +1,3 @@
 export * from './Input'
-export * from './Select'
-export * from './NativeSelect'
 export * from './TextField'
 export * from './types'
