+++ packages/ui/src/components/Input/Input/types.ts
@@ -8,11 +8,3 @@ export type InputType =
   | 'tel'
   | 'url'
   | 'number'
-
-export interface SelectItem {
-  value: string
-  label?: string
-  disabled?: boolean
-  group?: string
-  [key: string]: any
-}
