+++ packages/ui/src/components/Input/Select/NativeSelect.tsx
@@ -4,7 +4,7 @@ import { useUuid } from '@aviato/hooks'
 import { noop } from '@aviato/utils'
 
 import { StitchedCSS } from '~/theme'
-import { BaseInput, BaseInputProps, StyledInput } from './BaseInput'
+import { BaseInput, BaseInputProps, StyledInput } from '../Input/BaseInput'
 import { InputWrapper } from '../InputWrapper'
 import { SelectItem } from './types'
 import { onChange } from '~/types'
