+++ packages/ui/src/components/Input/Select/index.ts
@@ -1,0 +1,2 @@
+export * from './Select'
+export * from './NativeSelect'
