+++ packages/ui/src/components/Input/Select/types.ts
@@ -1,0 +1,7 @@
+export interface SelectItem {
+  value: string
+  label?: string
+  disabled?: boolean
+  group?: string
+  [key: string]: any
+}
